simplify brigadier logger; eliminate brigadier's third-party dependencies
