+++ v2/brigadier/package.json
@@ -34,8 +34,5 @@
     "mocha": "^8.2.1",
     "ts-node": "^9.1.1",
     "typescript": "^4.1.2"
-  },
-  "dependencies": {
-    "winston": "^3.3.3"
   }
 }
