+++ v2/brigadier/src/logger.ts
@@ -1,23 +1,25 @@
-import * as winston from "winston"
-import { TransformableInfo } from "logform"
+class Logger {
 
-export { Logger } from "winston" 
+  public debug(message: string, ...meta: any[]): Logger {
+    console.debug(message, ...meta)
+    return this
+  }
 
-export const logger = winston.createLogger({
-  levels: {
-    error: 0,
-    warn: 1,
-    info: 2,
-    debug: 3
-  },
-  format: winston.format.combine(
-    winston.format.timestamp(),
-    winston.format.printf((info: TransformableInfo) => {
-      if (info.job) {
-        return `${info.timestamp} [job: ${info.job}] ${info.level.toUpperCase()}: ${info.message}`
-      }
-      return `${info.timestamp} ${info.level.toUpperCase()}: ${info.message}`
-    })
-  ),
-  transports: [ new winston.transports.Console() ]
-})
+  public info(message: string, ...meta: any[]): Logger {
+    console.info(message, ...meta)
+    return this
+  }
+
+  public warn(message: string, ...meta: any[]): Logger {
+    console.warn(message, ...meta)
+    return this
+  }
+
+  public error(message: string, ...meta: any[]): Logger {
+    console.error(message, ...meta)
+    return this
+  }
+
+}
+
+export const logger = new Logger()
